fix component certificate
